Move CheckedConsumer, CheckedPredicate and CheckedRunnable from Try to io.vavr (#1968)
