+++ vavr/src/main/java/io/vavr/CheckedConsumer.java
@@ -1,0 +1,24 @@
+/*                        __    __  __  __    __  ___
+ *                       \  \  /  /    \  \  /  /  __/
+ *                        \  \/  /  /\  \  \/  /  /
+ *                         \____/__/  \__\____/__/.ɪᴏ
+ * ᶜᵒᵖʸʳᶦᵍʰᵗ ᵇʸ ᵛᵃᵛʳ ⁻ ˡᶦᶜᵉⁿˢᵉᵈ ᵘⁿᵈᵉʳ ᵗʰᵉ ᵃᵖᵃᶜʰᵉ ˡᶦᶜᵉⁿˢᵉ ᵛᵉʳˢᶦᵒⁿ ᵗʷᵒ ᵈᵒᵗ ᶻᵉʳᵒ
+ */
+package io.vavr;
+
+/**
+ * A {@linkplain java.util.function.Consumer} which may throw.
+ *
+ * @param <T> the type of value supplied to this consumer.
+ */
+@FunctionalInterface
+public interface CheckedConsumer<T> {
+
+    /**
+     * Performs side-effects.
+     *
+     * @param value a value
+     * @throws Throwable if an error occurs
+     */
+    void accept(T value) throws Throwable;
+}
