+++ vavr/src/main/java/io/vavr/CheckedPredicate.java
@@ -1,0 +1,34 @@
+/*                        __    __  __  __    __  ___
+ *                       \  \  /  /    \  \  /  /  __/
+ *                        \  \/  /  /\  \  \/  /  /
+ *                         \____/__/  \__\____/__/.ɪᴏ
+ * ᶜᵒᵖʸʳᶦᵍʰᵗ ᵇʸ ᵛᵃᵛʳ ⁻ ˡᶦᶜᵉⁿˢᵉᵈ ᵘⁿᵈᵉʳ ᵗʰᵉ ᵃᵖᵃᶜʰᵉ ˡᶦᶜᵉⁿˢᵉ ᵛᵉʳˢᶦᵒⁿ ᵗʷᵒ ᵈᵒᵗ ᶻᵉʳᵒ
+ */
+package io.vavr;
+
+/**
+ * A {@linkplain java.util.function.Predicate} which may throw.
+ *
+ * @param <T> the type of the input to the predicate
+ */
+@FunctionalInterface
+public interface CheckedPredicate<T> {
+
+    /**
+     * Evaluates this predicate on the given argument.
+     *
+     * @param t the input argument
+     * @return {@code true} if the input argument matches the predicate, otherwise {@code false}
+     * @throws Throwable if an error occurs
+     */
+    boolean test(T t) throws Throwable;
+
+    /**
+     * Negates this predicate.
+     *
+     * @return A new CheckedPredicate.
+     */
+    default CheckedPredicate<T> negate() {
+        return t -> !test(t);
+    }
+}
