+++ vavr/src/main/java/io/vavr/CheckedRunnable.java
@@ -1,0 +1,21 @@
+/*                        __    __  __  __    __  ___
+ *                       \  \  /  /    \  \  /  /  __/
+ *                        \  \/  /  /\  \  \/  /  /
+ *                         \____/__/  \__\____/__/.ɪᴏ
+ * ᶜᵒᵖʸʳᶦᵍʰᵗ ᵇʸ ᵛᵃᵛʳ ⁻ ˡᶦᶜᵉⁿˢᵉᵈ ᵘⁿᵈᵉʳ ᵗʰᵉ ᵃᵖᵃᶜʰᵉ ˡᶦᶜᵉⁿˢᵉ ᵛᵉʳˢᶦᵒⁿ ᵗʷᵒ ᵈᵒᵗ ᶻᵉʳᵒ
+ */
+package io.vavr;
+
+/**
+ * A {@linkplain Runnable} which may throw.
+ */
+@FunctionalInterface
+public interface CheckedRunnable {
+
+    /**
+     * Performs side-effects.
+     *
+     * @throws Throwable if an error occurs
+     */
+    void run() throws Throwable;
+}
